OmegaWiki TAGS
I created this so that the statistics on the Main Page via
Template:Statistics can automatically show the latest database count on
expression, dm and language. see
http://www.omegawiki.org/User:Hiong3.eng5/Statistics for the tag sample.

Kindly check for errors. If none, or if this feature is good, kindly
commit this, thanks.

P2: Used Cache. I also changed the tag format Note: I do not have cache. When I set up my MediaWiki/WikiLexicalData, there is this warning: Could not find APC, XCache or WinCache. Object caching is not enabled.

Therefore, can you check if it works. Thanks.

Change-Id: I7256dae2d0d505352c4988c3e1d66bf33cfec7e3