Remove Duplicate Expressions
Find, adjust, remove.

This does not delete the duplicate entry, just maked removed. In the
future, when the dm expression is not used, these duplicates should be
deleted. As a precaution, The reviveExpression function was changed to
limit the revival of expressions to 1 expression.

Some duplicates like Tala have the same transaction id, so some
expressions could be from the trailing space bug. Some may have been from
reviveEpression.  btw, Micheal's list must be rechecked in OmegaWiki
after this script is used. Like Tala is the same definition yet with
conflicting annotations.

I have checked SQL to see if the duplicates were deleted, checked if the
entries are ok, deleted all groundnut expression to see if the expression
is removed, then entered them again to see if the expression is revived.
The SQL said its all good, maybe you could double check this? Thanks.

Change-Id: Id70f8ed1ea392190d3b07f30815c6dfabf16c30a