Various db query function fix
converter's query adds the MySQL prefix and need not be fixed. Search.php
seems not used, if it is, I do not know how to test it.

Patch 2: language.php's query function need not be changed since the SQL
it uses was created using selectSQLText.

Query functions not yet checked are contained in
SpecialNeedTranslation.php, SpecialSuggest.php and WikiDataAPI.php.

There are PHP mysql_query functions at SpecialNeedsTranslation.php and
copytest.php and Copy.php which if I am correct, might need to be checked
for prefixed database. I am not sure, kindly check.

Kindly merge this if there are no more issues. Thanks.

Bug: 56220
Change-Id: Ia43d86734b2f59c39497946a236185c431c88700