Bug: 56220
Mostly documentation.

There are still query functions left.

There are those that are in unused files like Alert, SpecialTransaction,
copytest and update_bootstrap. These does not need to be changed for now,
since they are currently and may in the future be deleted, unless someone
uses these files.

Some are safe as is. RecordSetQueries and WikiDataAPI has one query
function each, but the query they execute are created by methods that
attach the database prefix, that is needed.

Then there is Copy.php. It uses php's mysql_query(), which means that
prefix is not attached by the software. Yet, only DefinedMeaningModel uses
Copy.php, and the only function it uses, CheckIfStub() is unused. This is
regarding the main program

A Special page, SpecialCopy, does use Copy.php, and if anyone is still
using this via SpecialPage Concept Mapping, then the mysql_query function
must be dealt with.  However, looking at Copy.php, I feel it is too
complex for me to fix.  If no one is using the Concept Mapping Page, Then
this bug can be labeled as fixed. If someone is going to use this special
page, then we will need to wait for someone else bold enough to fix this.
Sorry.

My contribution for this bug will only be up to this point. (In the
future, I may refactor RecordSetQueries and WikiDataAPI's query function,
and no more than that.)

Bug: 56220
Change-Id: I2a137d44786c0dec61c46ad20cb943097cd93e90