Install via MediaWiki's update.php
patch 2: The message to add globals on LocalSettings.php was not
noticable. So I created our own update.php that installs the base schema,
then calls MediaWiki's update.php class, and then display the message if
it is a new install. So we use update.php from WikiLexicalData's
maintenance folder. Updated documentation.

Change-Id: I2fddb4f8fcb1131559139ede8b3ec7f84951ba1c