Use getter instead of property
TextContent has the dedicated getter getNativeData() to access to the
object property mText, so let’s use it. I proposed to change the visibility
of the object property, so this extension would be broken if/when the
visibility is changed, but anyway it’s better to access through the getter.

Bug: T155298
Change-Id: I97a7c69a770844949bb4cf3de74b43dd85d4aaf6