Auto-fix many @param and @return PHPDoc tags
It seems this codebase uses abbreviations like "str" and "arr" for
types. I applied a bunch of regular expressions to auto-fix all PHPDoc
tags where I could identify a type.

Note this patch is intentionally "incomplete" in so far that it does
not fix all @param tags. More fixes will be done in later patches.

Change-Id: I411feea88a07b0be1fd94981884527e6e3a0cd69